requirements.txt
fastmcp>=0.2.0
azure-identity>=1.15.0
azure-mgmt-costmanagement>=4.0.0
azure-mgmt-consumption>=10.0.0
azure-mgmt-resource>=23.0.0
azure-mgmt-subscription>=3.1.0
azure-mgmt-advisor>=9.0.0
azure-mgmt-compute>=30.0.0
azure-mgmt-network>=25.0.0
cryptography>=42.0.0
pydantic>=2.5.0
python-dotenv>=1.0.0
flask>=3.0.0
